+++ apps/riscv-tui/src/event/key.rs
@@ -1,9 +1,10 @@
 use crossterm::event::KeyCode;
 
-use KeyControl::*;
+use NormalKeyControl::*;
+use EditKeyControl::*;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum KeyControl {
+pub enum NormalKeyControl {
     Quit,
     GoNext,
     GoPrev,
@@ -14,27 +15,58 @@ pub enum KeyControl {
     Step,
     RunToEnd,
     BreakPoint,
-    SearchMem,
+    SearchBus,
     ChangeView,
+    ShowInfo,
+    ClosePopup,
 }
 
-use KeyCode::*;
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum EditKeyControl {
+    Enter,
+    Delete,
+    CursorLeft,
+    CursorRight,
+    Exit,
+    Char(char),
+}
 
-pub fn poll_key_event(keycode: KeyCode) -> Option<KeyControl> {
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum KeyControl {
+    Normal(NormalKeyControl),
+    Edit(EditKeyControl),
+}
+
+pub fn get_normal_key(keycode: KeyCode) -> Option<NormalKeyControl> {
     Some(match keycode {
-        Char('q' | 'Q') => Quit,
-        Char('r' | 'R') => Reset,
-        Char('s' | 'S') => Step,
-        Char('p' | 'P') => RunToEnd,
-        #[cfg(feature = "zicsr")]
-        Char('c' | 'C') => ChangeMid,
-        Char('b' | 'B') => BreakPoint,
-        Char('m' | 'M') => SearchMem,
-        Char('h' | 'H') => ChangeView,
-        Up              => GoPrev,
-        Down            => GoNext,
-        Left | Right    => ChangePanel,
-        Tab             => ChangeMode,
-        _               => return None,
+            KeyCode::Char('q' | 'Q') => Quit,
+            KeyCode::Char('r' | 'R') => Reset,
+            KeyCode::Char('s' | 'S') => Step,
+            KeyCode::Char('p' | 'P') => RunToEnd,
+            #[cfg(feature = "zicsr")]
+            KeyCode::Char('c' | 'C') => ChangeMid,
+            KeyCode::Char('b' | 'B') => BreakPoint,
+            KeyCode::Char('v' | 'V') => SearchBus,
+            KeyCode::Char('h' | 'H') => ChangeView,
+            KeyCode::Char('i' | 'I') => ShowInfo,
+            KeyCode::Up              => GoPrev,
+            KeyCode::Down            => GoNext,
+            KeyCode::Left            => ChangePanel,
+            KeyCode::Right           => ChangePanel,
+            KeyCode::Tab             => ChangeMode,
+            KeyCode::Esc             => ClosePopup,
+            _                        => return None,
+    })
+}
+
+pub fn get_edit_key(keycode: KeyCode) -> Option<EditKeyControl> {
+    Some(match keycode {
+            KeyCode::Enter           => Enter,
+            KeyCode::Char(to_insert) => Char(to_insert),
+            KeyCode::Backspace       => Delete,
+            KeyCode::Left            => CursorLeft,
+            KeyCode::Right           => CursorRight,
+            KeyCode::Esc             => Exit,
+            _                        => return None,
     })
 }