+++ apps/riscv-tui/src/state.rs
@@ -4,20 +4,30 @@ mod snapshot;
 
 use std::collections::HashSet;
 
-use riscv_core::debug::DebugInterface;
+use riscv_core::debug::{DebugInterface, MachineInfo};
 
 use snapshot::MachineSnapshot;
 
 pub use mode::{DataView, EmuMode, Mid, Selected};
 
-#[derive(Debug, PartialEq, Eq)]
+use crate::input::EmuInput;
+
+#[derive(Debug, PartialEq)]
 pub struct EmuState {
     pub mach_snap: MachineSnapshot,
+    pub mach_info: MachineInfo,
 
     pub mode: EmuMode,
     pub selected: Selected,
     pub mid_selected: Mid,
     pub data_view: DataView,
+
+    pub show_search_popup: bool, 
+    pub show_info_popup: bool, 
+    pub show_bus_popup: bool, 
+    pub temp_bus_view: Option<(u32, Vec<u8>)>,
+
+    pub input: EmuInput,
 
     pub breakpoint_set: HashSet<usize>,
 }
@@ -25,18 +35,26 @@ pub struct EmuState {
 impl EmuState {
     pub fn new<D: DebugInterface>(mach: &D, ins_list: Vec<(u32, String)>) -> Self {
         let mach_snap = MachineSnapshot::new(mach, ins_list);
+        let mach_info = mach.get_info();
 
         let mode = EmuMode::default();
         let selected = Selected::default();
         let mid_selected = Mid::default();
         let data_view = DataView::default();
 
+        let show_search_popup = false;
+        let show_bus_popup = false;
+        let show_info_popup = false;
+        let temp_bus_view = None;
+        let input = EmuInput::default();
+
         let breakpoint_set = HashSet::new();
 
         EmuState { 
-            mach_snap,
+            mach_snap, mach_info,
             mode, selected, mid_selected, data_view,
-            breakpoint_set
+            show_search_popup, show_bus_popup, show_info_popup, 
+            temp_bus_view, input, breakpoint_set
         }
     }