+++ apps/riscv-tui/src/ui/component/mod.rs
@@ -2,11 +2,8 @@ mod instruction;
 mod exception;
 mod mid;
 
-use ratatui::{
-    Frame, 
-    layout::Rect,
-    style::{Color}
-};
+use ratatui::Frame;
+use ratatui::layout::Rect;
 
 use crate::state::EmuState;
 
@@ -15,10 +12,6 @@ pub use exception::Exception;
 #[cfg(feature = "zicsr")]
 pub use mid::csr::Csr;
 pub use mid::register::Register;
-
-pub(super) const ANTI_FLASH_WHITE: Color = Color::Rgb(242, 242, 242);
-pub(super) const BERKELEY_BLUE: Color = Color::Rgb(0, 50, 98);
-pub(super) const CALIFORNIA_GOLD: Color = Color::Rgb(253, 181, 21);
 
 pub trait Component {
     fn render(f: &mut Frame, area: Rect, emu: &mut EmuState);